+++ RELEASE-NOTES.md
@@ -14,6 +14,7 @@
 * #83 Convert Series to List, Set, array
 * #84 Shortcuts for Series aggregations 
 * #85 Series.map flavor that returns a DataFrame
+* #86 Expose JdbcConnector DataSource for ease of integration
 
 ## Release 0.6